Fix parsing build status with parentheses in name
Jenkins allows custom names for the builds.  If there were parentheses
in the name, the build status parsing would fail.  This corrects it
by only looking for the last pair of parentheses in the line, and falls
back to searching the whole string for the status, which handles cases
that include the previous successful build name as part of the status.